requirements.txt
feedparser==6.0.10
requests==2.31.0
pyyaml==6.0.1
flask==3.0.0
openai>=1.0.0
